refactor: Refactored NetSuite models
